Add logic for dev, and attempt to dev things

+++ snapshot/common.sh
@@ -24,8 +24,6 @@ function sanitize {
 function snapshot {
   # Remove initial '.'
   subdir=$(echo "${suffix}" | sed -e 's/^\.//')
-  # Remove _blacklist from JOB_NAME
-  JOB_NAME=$(echo $JOB_NAME | sed -e 's/_blacklist//')
   # Store reduce with dev, they are the same level of sanitization.
   if [ "${subdir}" = 'reduce' ]; then
     subdir='dev'
@@ -35,30 +33,33 @@ function snapshot {
   sudo innobackupex --no-timestamp /var/sanitize/drupal_export/${subdir}
   sudo innobackupex --apply-log --export "/var/sanitize/drupal_export/${subdir}"
   sudo chown -R bender:bender "/var/sanitize/drupal_export/${subdir}"
-  
-  # Ignore for now...
-  if [ 0 -eq 1 ]; then
-    tar -czvf "/var/dumps/${subdir}/${JOB_NAME}${suffix}-${BUILD_NUMBER}-binary.tar.gz" "/var/sanitize/drupal_export/${subdir}/${db}"
-    ln -sfv "${JOB_NAME}${suffix}-${BUILD_NUMBER}-binary.tar.gz" "/var/dumps/${subdir}/${JOB_NAME}${suffix}-binary-current.tar.gz"
-    # Don't forget me... remove old binary snapshots too.
-    old_snapshots=$(ls -t /var/dumps/${subdir}/${JOB_NAME}${suffix}-[0-9]*-binary.tar.gz | tail -n +2)
+  # Create a tarball for each database
+  for db in ${dblist}; do
+    mv "/var/sanitize/drupal_export/${db}${suffix}-schema.sql" "/var/sanitize/drupal_export/${subdir}/${db}"
+    cd "/var/sanitize/drupal_export/${subdir}/${db}"
+    tar -czvf "/var/dumps/${subdir}/${db}${suffix}-${BUILD_NUMBER}-binary.tar.gz" "./"
+    sudo chown -R bender:bender "/var/dumps/${subdir}/${db}${suffix}-${BUILD_NUMBER}-binary.tar.gz"
+    ln -sfv "${db}${suffix}-${BUILD_NUMBER}-binary.tar.gz" "/var/dumps/${subdir}/${db}${suffix}-binary-current.tar.gz"
+    # Remove old binary snapshots
+    old_snapshots=$(ls -t /var/dumps/${subdir}/${db}${suffix}-[0-9]*-binary.tar.gz | tail -n +2)
     if [ -n "${old_snapshots}" ]; then
       rm -v ${old_snapshots}
     fi
-    sudo rm -rf /var/sanitize/drupal_export/${subdir}/${db}
-    # Save the DB dump.
-    if [ "${subdir}" == 'dev' ]; then
-      sudo mysqldump --single-transaction --quick --max-allowed-packet=256M ${tmp_args} | sed -e 's/^) ENGINE=[^ ]*/) ROW_FORMAT=COMPRESSED/' | pbzip2 -p4 > "/var/dumps/${subdir}/${JOB_NAME}${suffix}-${BUILD_NUMBER}-in-progress.sql.bz2"
-      mv -v "/var/dumps/${subdir}/${JOB_NAME}${suffix}-${BUILD_NUMBER}-in-progress.sql.bz2" "/var/dumps/${subdir}/${JOB_NAME}${suffix}-${BUILD_NUMBER}.sql.bz2"
-      ln -sfv "${JOB_NAME}${suffix}-${BUILD_NUMBER}.sql.bz2" "/var/dumps/${subdir}/${JOB_NAME}${suffix}-current.sql.bz2"
-
+    # Also create old mysqldump snapshots for dev databases that aren't using
+    # the whitelist.
+    if [ "${subdir}" == 'dev' && "${db}" != 'drupal' ]; then
+      sudo mysqldump --single-transaction --quick --max-allowed-packet=256M ${db} | pbzip2 -p6 > "/var/dumps/${subdir}/${db}${suffix}-${BUILD_NUMBER}-in-progress.sql.bz2"
+      sudo chown -R bender:bender "/var/dumps/${subdir}/${db}${suffix}-${BUILD_NUMBER}-in-progress.sql.bz2"
+      mv -v "/var/dumps/${subdir}/${db}${suffix}-${BUILD_NUMBER}-in-progress.sql.bz2" "/var/dumps/${subdir}/${db}${suffix}-${BUILD_NUMBER}.sql.bz2"
+      ln -sfv "${db}${suffix}-${BUILD_NUMBER}.sql.bz2" "/var/dumps/${subdir}/${db}${suffix}-current.sql.bz2"
       # Remove old snapshots.
-      old_snapshots=$(ls -t /var/dumps/${subdir}/${JOB_NAME}${suffix}-[0-9]*.sql.{bz2,gz} | tail -n +2)
+      old_snapshots=$(ls -t /var/dumps/${subdir}/${db}${suffix}-[0-9]*.sql.{bz2,gz} | tail -n +2)
       if [ -n "${old_snapshots}" ]; then
         rm -v ${old_snapshots}
       fi
     fi
-  fi
+    cd "/usr/local/drupal-infrastructure"
+  done
 }
 
 function clear_tmp {
